+++ test/tests.lisp
@@ -80,7 +80,7 @@
   (let* ((boxed-string (box "Hello"))
          (ex (nth-value 1 (ignore-errors
                            (setf (ref boxed-string 0) #\A)))))
-    (is (typep ex 'accessor-resolution-error))))
+    (is (typep ex 'dotnet-error))))
 
 (test test-callback
   (let* ((x 1)